+++ src/main/java/dozun/game/threading/tasks/MultiThreadTask.java
@@ -18,9 +18,9 @@ public class MultiThreadTask implements Runnable {
 
     @Override
     public void run() {
-        Long dice1 = RandomNumberGenerator.getDice1();
-        Long dice2 = RandomNumberGenerator.getDice2();
-        Long dice3 = RandomNumberGenerator.getDice3();
+        int dice1 = RandomNumberGenerator.getDice1();
+        int dice2 = RandomNumberGenerator.getDice2();
+        int dice3 = RandomNumberGenerator.getDice3();
 
         diceResult.setDice1(dice1);
         diceResult.setDice1(dice2);
