+++ src/main/java/dozun/game/utils/GameGenerator.java
@@ -8,9 +8,9 @@ public class GameGenerator {
     private DiceResult diceResult;
 
     public DiceResult getGame(){
-        Long dice1 = RandomNumberGenerator.getDice1();
-        Long dice2 = RandomNumberGenerator.getDice2();
-        Long dice3 = RandomNumberGenerator.getDice3();
+        int dice1 = RandomNumberGenerator.getDice1();
+        int dice2 = RandomNumberGenerator.getDice2();
+        int dice3 = RandomNumberGenerator.getDice3();
 
         return new DiceResult(
                 dice1,
